requirements.txt
asgiref==3.4.1
boto3==1.20.6
botocore==1.23.6
certifi==2021.5.30
charset-normalizer==2.0.5
dj-database-url==0.5.0
Django==3.2.7
django-cors-headers==3.8.0
django-heroku==0.3.1
django-storages==1.12.3
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djongo==1.3.6
dnspython==2.1.0
gunicorn==20.1.0
idna==3.2
jmespath==0.10.0
Pillow==8.3.2
psycopg2==2.9.2
PyJWT==2.1.0
pymongo==3.12.1
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
s3transfer==0.5.0
six==1.16.0
sqlparse==0.2.4
stripe==2.60.0
urllib3==1.26.6
whitenoise==5.3.0
